[Feat.] Add function to update an acl rule
